refactoring
Some refactoring thanks to the new structure
