+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/controller/AuthenticationController.kt
@@ -1,7 +1,7 @@
 package com.vitorbasso.gerenciadorinvestimentos.controller
 
 import com.vitorbasso.gerenciadorinvestimentos.dto.request.AuthenticationRequest
-import com.vitorbasso.gerenciadorinvestimentos.service.concrete.AuthenticationService
+import com.vitorbasso.gerenciadorinvestimentos.service.security.AuthenticationService
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
