+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/repository/WalletRepository.kt
@@ -1,12 +1,13 @@
 package com.vitorbasso.gerenciadorinvestimentos.repository
 
+import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Client
 import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Wallet
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
 interface WalletRepository : JpaRepository<Wallet, Int> {
-    fun findByBrokerAndClientId(broker: String, clientId: Long): Wallet?
+    fun findByBrokerAndClient(broker: String, client: Client): Wallet?
 
-    fun existsByBrokerAndClientId(broker: String, clientId: Long): Boolean
+    fun existsByBrokerAndClient(broker: String, client: Client): Boolean
 }
