+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/service/concrete/ClientService.kt
@@ -24,7 +24,7 @@ internal class ClientService (
             = clientRepository.findByIdOrNull(id) ?: throw exception
 
     fun saveClient(clientToSave: Client)
-            = if(!exists(clientToSave.cpf))
+            = if(!exists(clientToSave.id))
         this.clientRepository.save(clientToSave.copy(password = this.passwordEncoder.encode(clientToSave.password)))
         else throw CustomBadRequestException(ManagerErrorCode.MANAGER_04)
 
@@ -39,6 +39,6 @@ internal class ClientService (
         this.clientRepository.delete(clientToDelete)
     }
 
-    private fun exists(cpf: String) = this.clientRepository.existsByCpf(cpf)
+    private fun exists(id: Long) = this.clientRepository.existsById(id)
 
 }
